feat(resultsAppExecConnect): Abstracted details page header into a separate component
