+++ src/Analysis/GWASResultsV2/SharedComponents/DetailPageHeader/DetailPageHeader.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import ReturnHomeButton from './ReturnHomeButton/ReturnHomeButton';
+import './DetailPageHeader.css';
+
+const DetailPageHeader = ({ PageTitle }) => (
+  <div className='details-page-header'>
+    <ReturnHomeButton />
+    <h1>{PageTitle}</h1>
+  </div>
+);
+export default DetailPageHeader;
